+++ multidla/counter.cpp
@@ -22,9 +22,9 @@ void Counter::next()
 {
     m_isNext = false;
     for (int32_t d = int32_t(m_current.defDims()) - 1; d >= 0; --d) {
-        m_current.setCoord(d, m_current.coord(d) + 1);
-        if (m_current.coord(d) == m_size.coord(d)) {
-            m_current.setCoord(d, 0);
+        m_current.setCoord(size_t(d), m_current.coord(size_t(d)) + 1);
+        if (m_current.coord(size_t(d)) == m_size.coord(size_t(d))) {
+            m_current.setCoord(size_t(d), 0);
         } else {
             m_isNext = true;
             break;
